// src/app/model/loan-account.model.ts
export class LoanAccount {
    accountId: string;
    firstName: string;
    lastName: string;
    citizenId: string;
    bankName: string;
    baneAccountName: string;
    baneAccountId: string;
    telMain: string;
    tel1: string;
    tel2: string;
    lineId: string;
    email: string;
    address: string;
    career: string;
    salary: string;
    relationshipAccountId: string;
    relationshipDetail: string;
    picture: string;
    other: string;
}